Introduce HttpSourceBase
