+++ mycroft_holmes/sources/http/__init__.py
@@ -1,0 +1,4 @@
+"""
+HTTP-based sources
+"""
+from .base import HttpSourceBase
